Extract plugin parser stuff into separate module
